feat: 측정 결과 저장 기능 구현

+++ Android/app/src/main/java/com/example/myapplication/data/dto/model/BloodPressureResult.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.Serializable
 data class BloodPressureResult(
     @SerialName("FitrusDevice") val device: String,
     @SerialName("FirmwareVersion") val firmware: String,
-    val dbp: Double,
-    val sbp: Double,
-    val baseSBP: Double,
-    val baseDBP: Double
+    val dbp: Float,
+    val sbp: Float,
+    val baseSBP: Float,
+    val baseDBP: Float
 ) : MeasureResult()
