+++ Android/app/src/main/java/com/example/myapplication/data/dto/model/StressResult.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.Serializable
 data class StressResult(
     @SerialName("FitrusDevice") val device: String,
     @SerialName("FirmwareVersion") val firmware: String,
-    val value: Int,
+    val value: Short,
     val level: String,
-    val oxygen: Int,
-    val bpm: Int
+    val oxygen: Short,
+    val bpm: Short
 ) : MeasureResult()
